Add builder module
